+++ src/main/java/com/sbt/concurrent/MutexCounter.java
@@ -1,7 +1,7 @@
 package com.sbt.concurrent;
 
 public class MutexCounter implements Counter{
-    private int i = 0;
+    private long i = 0;
 
     public synchronized void increment() {
         i++;
